requirements.txt
streamlit
google-genai
feedparser
beautifulsoup4
PyGithub
python-dotenv
